requirements.txt
PyQt6
pdf2image
Pillow
opencv-python
pandas
pytesseract
pymupdf